refactor: Move GetByIdRequestDTO to request-dto directory
