+++ host-frontend-root/frontend-src-root/src/frameworks-and-drivers/messaging/dto/request-dto/GetByIdRequestDTO.ts
