+++ host-frontend-root/frontend-src-root/src/interface-adapters/ports/IRewriteRuleMessagingPort.ts
@@ -1,4 +1,4 @@
-import { GetByIdRequestDTO } from 'src/frameworks-and-drivers/messaging/dto/GetByIdRequestDTO';
+import { GetByIdRequestDTO } from 'src/frameworks-and-drivers/messaging/dto/request-dto/GetByIdRequestDTO';
 import { UpdateRuleActiveRequestDTO } from 'src/frameworks-and-drivers/messaging/dto/request-dto/UpdateRuleActiveRequestDTO';
 import { RewriteRuleDTO } from 'src/frameworks-and-drivers/messaging/dto/RewriteRuleDTO';
